ml_service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
Pillow
huggingface_hub
python-multipart